Naively replace of size_t with uint64_t
This is dummy.
Just as reference to show it works

+++ src/bindings/jni/rvvm_jni.c
@@ -42,7 +42,7 @@ JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_create_1machine(JNIEnv* env,
 {
     const char* u8_isa = (*env)->GetStringUTFChars(env, isa, NULL);
     UNUSED(class);
-    jlong ret = (size_t)rvvm_create_machine(mem_size, smp, u8_isa);
+    jlong ret = (uint64_t)rvvm_create_machine(mem_size, smp, u8_isa);
     (*env)->ReleaseStringUTFChars(env, isa, u8_isa);
     return ret;
 }
@@ -51,7 +51,7 @@ JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_set_1cmdline(JNIEnv* env, jcl
 {
     const char* u8_cmdline = (*env)->GetStringUTFChars(env, cmdline, NULL);
     UNUSED(class);
-    rvvm_set_cmdline((rvvm_machine_t*)(size_t)machine, u8_cmdline);
+    rvvm_set_cmdline((rvvm_machine_t*)(uint64_t)machine, u8_cmdline);
     (*env)->ReleaseStringUTFChars(env, cmdline, u8_cmdline);
 }
 
@@ -59,14 +59,14 @@ JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_append_1cmdline(JNIEnv* env, 
 {
     const char* u8_cmdline = (*env)->GetStringUTFChars(env, cmdline, NULL);
     UNUSED(class);
-    rvvm_append_cmdline((rvvm_machine_t*)(size_t)machine, u8_cmdline);
+    rvvm_append_cmdline((rvvm_machine_t*)(uint64_t)machine, u8_cmdline);
     (*env)->ReleaseStringUTFChars(env, cmdline, u8_cmdline);
 }
 
 JNIEXPORT jboolean JNICALL Java_lekkit_rvvm_RVVMNative_load_1bootrom(JNIEnv* env, jclass class, jlong machine, jstring path)
 {
     const char* u8_path = (*env)->GetStringUTFChars(env, path, NULL);
-    bool ret = rvvm_load_bootrom((rvvm_machine_t*)(size_t)machine, u8_path);
+    bool ret = rvvm_load_bootrom((rvvm_machine_t*)(uint64_t)machine, u8_path);
     UNUSED(class);
     (*env)->ReleaseStringUTFChars(env, path, u8_path);
     return ret;
@@ -75,7 +75,7 @@ JNIEXPORT jboolean JNICALL Java_lekkit_rvvm_RVVMNative_load_1bootrom(JNIEnv* env
 JNIEXPORT jboolean JNICALL Java_lekkit_rvvm_RVVMNative_load_1kernel(JNIEnv* env, jclass class, jlong machine, jstring path)
 {
     const char* u8_path = (*env)->GetStringUTFChars(env, path, NULL);
-    bool ret = rvvm_load_kernel((rvvm_machine_t*)(size_t)machine, u8_path);
+    bool ret = rvvm_load_kernel((rvvm_machine_t*)(uint64_t)machine, u8_path);
     UNUSED(class);
     (*env)->ReleaseStringUTFChars(env, path, u8_path);
     return ret;
@@ -84,7 +84,7 @@ JNIEXPORT jboolean JNICALL Java_lekkit_rvvm_RVVMNative_load_1kernel(JNIEnv* env,
 JNIEXPORT jboolean JNICALL Java_lekkit_rvvm_RVVMNative_load_1dtb(JNIEnv* env, jclass class, jlong machine, jstring path)
 {
     const char* u8_path = (*env)->GetStringUTFChars(env, path, NULL);
-    bool ret = rvvm_load_dtb((rvvm_machine_t*)(size_t)machine, u8_path);
+    bool ret = rvvm_load_dtb((rvvm_machine_t*)(uint64_t)machine, u8_path);
     UNUSED(class);
     (*env)->ReleaseStringUTFChars(env, path, u8_path);
     return ret;
@@ -93,7 +93,7 @@ JNIEXPORT jboolean JNICALL Java_lekkit_rvvm_RVVMNative_load_1dtb(JNIEnv* env, jc
 JNIEXPORT jboolean JNICALL Java_lekkit_rvvm_RVVMNative_dump_1dtb(JNIEnv* env, jclass class, jlong machine, jstring path)
 {
     const char* u8_path = (*env)->GetStringUTFChars(env, path, NULL);
-    bool ret = rvvm_dump_dtb((rvvm_machine_t*)(size_t)machine, u8_path);
+    bool ret = rvvm_dump_dtb((rvvm_machine_t*)(uint64_t)machine, u8_path);
     UNUSED(class);
     (*env)->ReleaseStringUTFChars(env, path, u8_path);
     return ret;
@@ -102,50 +102,50 @@ JNIEXPORT jboolean JNICALL Java_lekkit_rvvm_RVVMNative_dump_1dtb(JNIEnv* env, jc
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_get_1opt(JNIEnv* env, jclass class, jlong machine, jint opt)
 {
     UNUSED(env); UNUSED(class);
-    return rvvm_get_opt((rvvm_machine_t*)(size_t)machine, (uint32_t)opt);
+    return rvvm_get_opt((rvvm_machine_t*)(uint64_t)machine, (uint32_t)opt);
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_set_1opt(JNIEnv* env, jclass class, jlong machine, jint opt, jlong val)
 {
     UNUSED(env); UNUSED(class);
-    rvvm_set_opt((rvvm_machine_t*)(size_t)machine, (uint32_t)opt, val);
+    rvvm_set_opt((rvvm_machine_t*)(uint64_t)machine, (uint32_t)opt, val);
 }
 
 JNIEXPORT jboolean JNICALL Java_lekkit_rvvm_RVVMNative_start_1machine(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    return rvvm_start_machine((rvvm_machine_t*)(size_t)machine);
+    return rvvm_start_machine((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT jboolean JNICALL Java_lekkit_rvvm_RVVMNative_pause_1machine(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    return rvvm_pause_machine((rvvm_machine_t*)(size_t)machine);
+    return rvvm_pause_machine((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT jboolean JNICALL Java_lekkit_rvvm_RVVMNative_reset_1machine(JNIEnv* env, jclass class, jlong machine, jboolean reset)
 {
     UNUSED(env); UNUSED(class);
-    rvvm_reset_machine((rvvm_machine_t*)(size_t)machine, reset);
+    rvvm_reset_machine((rvvm_machine_t*)(uint64_t)machine, reset);
     return true;
 }
 
 JNIEXPORT jboolean JNICALL Java_lekkit_rvvm_RVVMNative_machine_1running(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    return rvvm_machine_running((rvvm_machine_t*)(size_t)machine);
+    return rvvm_machine_running((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT jboolean JNICALL Java_lekkit_rvvm_RVVMNative_machine_1powered(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    return rvvm_machine_powered((rvvm_machine_t*)(size_t)machine);
+    return rvvm_machine_powered((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_free_1machine(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    rvvm_free_machine((rvvm_machine_t*)(size_t)machine);
+    rvvm_free_machine((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_run_1eventloop(JNIEnv* env, jclass class)
@@ -160,7 +160,7 @@ JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_run_1eventloop(JNIEnv* env, j
 
 JNIEXPORT jobject JNICALL Java_lekkit_rvvm_RVVMNative_get_1dma_1buf(JNIEnv* env, jclass class, jlong machine, jlong addr, jlong size)
 {
-    void* ptr = rvvm_get_dma_ptr((rvvm_machine_t*)(size_t)machine, addr, size);
+    void* ptr = rvvm_get_dma_ptr((rvvm_machine_t*)(uint64_t)machine, addr, size);
     UNUSED(class);
     if (ptr == NULL) return NULL;
     return (*env)->NewDirectByteBuffer(env, ptr, size);
@@ -169,49 +169,49 @@ JNIEXPORT jobject JNICALL Java_lekkit_rvvm_RVVMNative_get_1dma_1buf(JNIEnv* env,
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_mmio_1zone_1auto(JNIEnv* env, jclass class, jlong machine, jlong addr, jlong size)
 {
     UNUSED(env); UNUSED(class);
-    return rvvm_mmio_zone_auto((rvvm_machine_t*)(size_t)machine, addr, size);
+    return rvvm_mmio_zone_auto((rvvm_machine_t*)(uint64_t)machine, addr, size);
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_remove_1mmio(JNIEnv* env, jclass class, jlong mmio_dev)
 {
     UNUSED(env); UNUSED(class);
-    rvvm_remove_mmio((rvvm_mmio_dev_t*)(size_t)mmio_dev);
+    rvvm_remove_mmio((rvvm_mmio_dev_t*)(uint64_t)mmio_dev);
 }
 
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_get_1intc(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    return (size_t)rvvm_get_intc((rvvm_machine_t*)(size_t)machine);
+    return (uint64_t)rvvm_get_intc((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_set_1intc(JNIEnv* env, jclass class, jlong machine, jlong intc)
 {
     UNUSED(env); UNUSED(class);
-    rvvm_set_intc((rvvm_machine_t*)(size_t)machine, (rvvm_intc_t*)(size_t)intc);
+    rvvm_set_intc((rvvm_machine_t*)(uint64_t)machine, (rvvm_intc_t*)(uint64_t)intc);
 }
 
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_get_1pci_1bus(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    return (size_t)rvvm_get_pci_bus((rvvm_machine_t*)(size_t)machine);
+    return (uint64_t)rvvm_get_pci_bus((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_set_1pci_1bus(JNIEnv* env, jclass class, jlong machine, jlong pci_bus)
 {
     UNUSED(env); UNUSED(class);
-    rvvm_set_pci_bus((rvvm_machine_t*)(size_t)machine, (pci_bus_t*)(size_t)pci_bus);
+    rvvm_set_pci_bus((rvvm_machine_t*)(uint64_t)machine, (pci_bus_t*)(uint64_t)pci_bus);
 }
 
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_get_1i2c_1bus(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    return (size_t)rvvm_get_i2c_bus((rvvm_machine_t*)(size_t)machine);
+    return (uint64_t)rvvm_get_i2c_bus((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_set_1i2c_1bus(JNIEnv* env, jclass class, jlong machine, jlong i2c_bus)
 {
     UNUSED(env); UNUSED(class);
-    rvvm_set_i2c_bus((rvvm_machine_t*)(size_t)machine, (i2c_bus_t*)(size_t)i2c_bus);
+    rvvm_set_i2c_bus((rvvm_machine_t*)(uint64_t)machine, (i2c_bus_t*)(uint64_t)i2c_bus);
 }
 
 /*
@@ -221,82 +221,82 @@ JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_set_1i2c_1bus(JNIEnv* env, jc
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_riscv_1clint_1init_1auto(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    riscv_clint_init_auto((rvvm_machine_t*)(size_t)machine);
+    riscv_clint_init_auto((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_riscv_1imsic_1init_1auto(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    riscv_imsic_init_auto((rvvm_machine_t*)(size_t)machine);
+    riscv_imsic_init_auto((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_riscv_1plic_1init_1auto(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    return (size_t)riscv_plic_init_auto((rvvm_machine_t*)(size_t)machine);
+    return (uint64_t)riscv_plic_init_auto((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_riscv_1aplic_1init_1auto(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    return (size_t)riscv_aplic_init_auto((rvvm_machine_t*)(size_t)machine);
+    return (uint64_t)riscv_aplic_init_auto((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_pci_1bus_1init_1auto(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    return (size_t)pci_bus_init_auto((rvvm_machine_t*)(size_t)machine);
+    return (uint64_t)pci_bus_init_auto((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_i2c_1bus_1init_1auto(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    return (size_t)i2c_oc_init_auto((rvvm_machine_t*)(size_t)machine);
+    return (uint64_t)i2c_oc_init_auto((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_tap_1user_1open(JNIEnv* env, jclass class)
 {
     UNUSED(env); UNUSED(class);
-    return (size_t)tap_open();
+    return (uint64_t)tap_open();
 }
 
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_syscon_1init_1auto(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    return (size_t)syscon_init_auto((rvvm_machine_t*)(size_t)machine);
+    return (uint64_t)syscon_init_auto((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_rtc_1goldfish_1init_1auto(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    return (size_t)rtc_goldfish_init_auto((rvvm_machine_t*)(size_t)machine);
+    return (uint64_t)rtc_goldfish_init_auto((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_rtc_1ds1742_1init_1auto(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    return (size_t)rtc_ds1742_init_auto((rvvm_machine_t*)(size_t)machine);
+    return (uint64_t)rtc_ds1742_init_auto((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_ns16550a_1init_1auto(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    return (size_t)ns16550a_init_term_auto((rvvm_machine_t*)(size_t)machine);
+    return (uint64_t)ns16550a_init_term_auto((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_gpio_1sifive_1init_1auto(JNIEnv* env, jclass class, jlong machine, jlong gpio)
 {
     UNUSED(env); UNUSED(class);
-    return (size_t)gpio_sifive_init_auto((rvvm_machine_t*)(size_t)machine, (rvvm_gpio_dev_t*)(size_t)gpio);
+    return (uint64_t)gpio_sifive_init_auto((rvvm_machine_t*)(uint64_t)machine, (rvvm_gpio_dev_t*)(uint64_t)gpio);
 }
 
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_mtd_1physmap_1init_1auto(JNIEnv* env, jclass class, jlong machine, jstring path, jboolean rw)
 {
     const char* u8_path = (*env)->GetStringUTFChars(env, path, NULL);
-    rvvm_mmio_dev_t* mmio = mtd_physmap_init_auto((rvvm_machine_t*)(size_t)machine, u8_path, rw);
+    rvvm_mmio_dev_t* mmio = mtd_physmap_init_auto((rvvm_machine_t*)(uint64_t)machine, u8_path, rw);
     UNUSED(class);
     (*env)->ReleaseStringUTFChars(env, path, u8_path);
-    return (size_t)mmio;
+    return (uint64_t)mmio;
 }
 
 static void jni_framebuffer_remove(rvvm_mmio_dev_t* dev)
@@ -331,7 +331,7 @@ JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_framebuffer_1init_1auto(JNIE
         return 0;
     }
 
-    rvvm_mmio_dev_t* mmio = framebuffer_init_auto((rvvm_machine_t*)(size_t)machine, &fb_ctx);
+    rvvm_mmio_dev_t* mmio = framebuffer_init_auto((rvvm_machine_t*)(uint64_t)machine, &fb_ctx);
     if (mmio) {
         // Return direct ByteBuffer to Java side, register framebuffer cleanup callback
         jobject bytebuf = (*env)->NewDirectByteBuffer(env, fb_ctx.buffer, framebuffer_size(&fb_ctx));
@@ -343,7 +343,7 @@ JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_framebuffer_1init_1auto(JNIE
         }
 
         (*env)->SetObjectArrayElement(env, fb, 0, bytebuf);
-        return (size_t)mmio;
+        return (uint64_t)mmio;
     }
 
     return 0;
@@ -353,28 +353,28 @@ JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_framebuffer_1init_1auto(JNIE
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_rtl8169_1init(JNIEnv* env, jclass class, jlong pci_bus, jlong tap)
 {
     UNUSED(env); UNUSED(class);
-    return (size_t)rtl8169_init((pci_bus_t*)(size_t)pci_bus, (tap_dev_t*)(size_t)tap);
+    return (uint64_t)rtl8169_init((pci_bus_t*)(uint64_t)pci_bus, (tap_dev_t*)(uint64_t)tap);
 }
 
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_nvme_1init(JNIEnv* env, jclass class, jlong pci_bus, jstring path, jboolean rw)
 {
     const char* u8_path = (*env)->GetStringUTFChars(env, path, NULL);
-    pci_dev_t* ret = nvme_init((pci_bus_t*)(size_t)pci_bus, u8_path, rw);
+    pci_dev_t* ret = nvme_init((pci_bus_t*)(uint64_t)pci_bus, u8_path, rw);
     UNUSED(class);
     (*env)->ReleaseStringUTFChars(env, path, u8_path);
-    return (size_t)ret;
+    return (uint64_t)ret;
 }
 
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_hid_1mouse_1init_1auto(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    return (size_t)hid_mouse_init_auto((rvvm_machine_t*)(size_t)machine);
+    return (uint64_t)hid_mouse_init_auto((rvvm_machine_t*)(uint64_t)machine);
 }
 
 JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_hid_1keyboard_1init_1auto(JNIEnv* env, jclass class, jlong machine)
 {
     UNUSED(env); UNUSED(class);
-    return (size_t)hid_keyboard_init_auto((rvvm_machine_t*)(size_t)machine);
+    return (uint64_t)hid_keyboard_init_auto((rvvm_machine_t*)(uint64_t)machine);
 }
 
 static void jni_gpio_remove(rvvm_gpio_dev_t* gpio)
@@ -387,18 +387,18 @@ JNIEXPORT jlong JNICALL Java_lekkit_rvvm_RVVMNative_gpio_1dev_1create(JNIEnv* en
     rvvm_gpio_dev_t* gpio = safe_new_obj(rvvm_gpio_dev_t);
     gpio->remove = jni_gpio_remove;
     UNUSED(env); UNUSED(class);
-    return (size_t)gpio;
+    return (uint64_t)gpio;
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_pci_1remove_1device(JNIEnv* env, jclass class, jlong pci_dev)
 {
     UNUSED(env); UNUSED(class);
-    pci_remove_device((pci_dev_t*)(size_t)pci_dev);
+    pci_remove_device((pci_dev_t*)(uint64_t)pci_dev);
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_gpio_1dev_1free(JNIEnv* env, jclass class, jlong gpio)
 {
-    void* ptr = (void*)(size_t)gpio;
+    void* ptr = (void*)(uint64_t)gpio;
     UNUSED(env); UNUSED(class);
     free(ptr);
 }
@@ -406,61 +406,61 @@ JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_gpio_1dev_1free(JNIEnv* env, 
 JNIEXPORT jint JNICALL Java_lekkit_rvvm_RVVMNative_gpio_1read_1pins(JNIEnv* env, jclass class, jlong gpio, jint off)
 {
     UNUSED(env); UNUSED(class);
-    return gpio_read_pins((rvvm_gpio_dev_t*)(size_t)gpio, off);
+    return gpio_read_pins((rvvm_gpio_dev_t*)(uint64_t)gpio, off);
 }
 
 JNIEXPORT jboolean JNICALL Java_lekkit_rvvm_RVVMNative_gpio_1write_1pins(JNIEnv* env, jclass class, jlong gpio, jint off, jint pins)
 {
     UNUSED(env); UNUSED(class);
-    return gpio_write_pins((rvvm_gpio_dev_t*)(size_t)gpio, off, pins);
+    return gpio_write_pins((rvvm_gpio_dev_t*)(uint64_t)gpio, off, pins);
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_hid_1mouse_1resolution(JNIEnv* env, jclass class, jlong mice, jint x, jint y)
 {
     UNUSED(env); UNUSED(class);
-    hid_mouse_resolution((hid_mouse_t*)(size_t)mice, x, y);
+    hid_mouse_resolution((hid_mouse_t*)(uint64_t)mice, x, y);
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_hid_1mouse_1place(JNIEnv* env, jclass class, jlong mice, jint x, jint y)
 {
     UNUSED(env); UNUSED(class);
-    hid_mouse_place((hid_mouse_t*)(size_t)mice, x, y);
+    hid_mouse_place((hid_mouse_t*)(uint64_t)mice, x, y);
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_hid_1mouse_1move(JNIEnv* env, jclass class, jlong mice, jint x, jint y)
 {
     UNUSED(env); UNUSED(class);
-    hid_mouse_move((hid_mouse_t*)(size_t)mice, x, y);
+    hid_mouse_move((hid_mouse_t*)(uint64_t)mice, x, y);
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_hid_1mouse_1press(JNIEnv* env, jclass class, jlong mice, jbyte btns)
 {
     UNUSED(env); UNUSED(class);
-    hid_mouse_press((hid_mouse_t*)(size_t)mice, btns);
+    hid_mouse_press((hid_mouse_t*)(uint64_t)mice, btns);
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_hid_1mouse_1release(JNIEnv* env, jclass class, jlong mice, jbyte btns)
 {
     UNUSED(env); UNUSED(class);
-    hid_mouse_release((hid_mouse_t*)(size_t)mice, btns);
+    hid_mouse_release((hid_mouse_t*)(uint64_t)mice, btns);
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_hid_1mouse_1scroll(JNIEnv* env, jclass class, jlong mice, jint offset)
 {
     UNUSED(env); UNUSED(class);
-    hid_mouse_scroll((hid_mouse_t*)(size_t)mice, offset);
+    hid_mouse_scroll((hid_mouse_t*)(uint64_t)mice, offset);
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_hid_1keyboard_1press(JNIEnv* env, jclass class, jlong kb, jbyte key)
 {
     UNUSED(env); UNUSED(class);
-    hid_keyboard_press((hid_keyboard_t*)(size_t)kb, key);
+    hid_keyboard_press((hid_keyboard_t*)(uint64_t)kb, key);
 }
 
 JNIEXPORT void JNICALL Java_lekkit_rvvm_RVVMNative_hid_1keyboard_1release(JNIEnv* env, jclass class, jlong kb, jbyte key)
 {
     UNUSED(env); UNUSED(class);
-    hid_keyboard_release((hid_keyboard_t*)(size_t)kb, key);
+    hid_keyboard_release((hid_keyboard_t*)(uint64_t)kb, key);
 }
 
 POP_OPTIMIZATION_SIZE
